fix(iterable): always encode chunks
