+++ test/_setup.ts
@@ -35,8 +35,12 @@ export function describeMatrix(
     fn(ctx, utils);
   };
   describe(title, () => {
-    run(setupWebTest(opts));
-    run(setupNodeTest(opts));
+    describe("web", () => {
+      run(setupWebTest(opts));
+    });
+    describe("node", () => {
+      run(setupNodeTest(opts));
+    });
   });
 }
